Refactor: Move manual function to dedicated file
