+++ include/manual.h
@@ -1,0 +1,6 @@
+#ifndef MANUAL_H
+#define MANUAL_H
+
+void manual(int page);
+
+#endif
